+++ src/features/export-package-resources-modal/utils.test.js
@@ -6,14 +6,14 @@ describe('ExportPackageResourcesModal utils', () => {
     describe('when all fields are defined in a constant', () => {
       it('should return correct mapped values', () => {
         expect(utils.formatExportFieldsPayload(['providerName', 'packageAgreements'], RECORD_TYPES.PACKAGE))
-          .toEqual(['providerName', 'packageAgreementStartDate', 'packageAgreementName', 'packageAgreementStatus']);
+          .toEqual(['providerName', 'packageAgreements']);
       });
     });
 
     describe('when some fields are not defined in a constant', () => {
       it('should return correct mapped values', () => {
         expect(utils.formatExportFieldsPayload(['providerName', 'nonExistingField', 'packageAgreements'], RECORD_TYPES.PACKAGE))
-          .toEqual(['providerName', 'packageAgreementStartDate', 'packageAgreementName', 'packageAgreementStatus']);
+          .toEqual(['providerName', 'packageAgreements']);
       });
     });
   });